requirements.txt
numpy==1.22.1
opencv-python==4.5.5.62
pandas==1.4.0
Pillow==9.0.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.0.1
